Add a New List of Favorite Songs
The Subject Line above conforms to the following standards:

(1) In the imperative mood: "If applied, this commit will _______"
(2) The words are in title case
(3) It is less than or equal to 50 characters long.

This is the body, which provides a description. These aren't really my
favorite songs. They are catchy songs that I'm ashamed of liking, except
Nickle Back I don't even really like this song but I think it's funny to
pretend that I do.

Here is a fake JIRA ticket: #123