Add a New List of Favorite Books
The Subject Line above conforms to the following standards:

(1) In the imperative mood: "If applied, this commit will _______"
(2) The words are in title case
(3) It is less than or equal to 50 characters long.

This is the body, which provides a description. These actually are some
of my favorite books. Plus A Tale of Two Cities for Ants.

Here is a fake JIRA ticket: #12345